dash-app/requirements.txt
dash==1.10.0
requests
folium
watson-machine-learning-client-V4
python-dotenv
